Remove old login components, reorganize login modules.
