--- client/src/component/login/LoginFormContainer.tsx
@@ -1,29 +1,0 @@
-import * as React from 'react';
-import { connect } from 'react-redux';
-
-import LoginForm, { FormFields } from './LoginForm';
-
-import countyLogin from 'corla/action/countyLogin';
-import dosLogin from 'corla/action/dosLogin';
-
-
-const submit = ({ dashboard, username, password }: any) => {
-    switch (dashboard) {
-        case 'sos': {
-            dosLogin(username, password);
-            break;
-        }
-        case 'county': {
-            countyLogin(username, password);
-            break;
-        }
-    }
-};
-
-export class LoginFormContainer extends React.Component<any, any> {
-    public render() {
-        return <LoginForm submit={ submit } />;
-    }
-}
-
-export default connect()(LoginFormContainer);
